+++ cmd/zetae2e/local/config.go
@@ -1,0 +1,37 @@
+package local
+
+import (
+	"path/filepath"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/spf13/cobra"
+	"github.com/zeta-chain/zetacore/app"
+	"github.com/zeta-chain/zetacore/e2e/config"
+)
+
+// GetConfig returns config from file from the command line flag
+func GetConfig(cmd *cobra.Command) (config.Config, error) {
+	configFile, err := cmd.Flags().GetString(FlagConfigFile)
+	if err != nil {
+		return config.Config{}, err
+	}
+
+	// use default config if no config file is specified
+	if configFile == "" {
+		return config.DefaultConfig(), nil
+	}
+
+	configFile, err = filepath.Abs(configFile)
+	if err != nil {
+		return config.Config{}, err
+	}
+
+	return config.ReadConfig(configFile)
+}
+
+// setCosmosConfig set account prefix to zeta
+func setCosmosConfig() {
+	cosmosConf := sdk.GetConfig()
+	cosmosConf.SetBech32PrefixForAccount(app.Bech32PrefixAccAddr, app.Bech32PrefixAccPub)
+	cosmosConf.Seal()
+}